+++ wgscovplot/db.py
@@ -1,5 +1,3 @@
-from typing import Dict, List
-
 from pydantic import BaseModel
 
 from wgscovplot.features import EChartsFeature
@@ -8,7 +6,7 @@ from wgscovplot.tools.mosdepth.flu import FluMosdepthDepthInfo
 
 
 class DBMixin(BaseModel):
-    samples: List[str]
+    samples: list[str]
     low_coverage_threshold: int = 10
 
 
@@ -18,13 +16,13 @@ class SegmentedGenomeDB(DBMixin):
     The single instance of this class will be serialized into a JS object for populating plots in the final HTML output.
     """
 
-    segments: List[str]
-    segments_ref_id: Dict[str, Dict[str, str]]
-    segments_ref_seq: Dict[str, Dict[str, str]]
-    depths: Dict[str, Dict[str, str]]
-    mosdepth_info: Dict[str, Dict[str, FluMosdepthDepthInfo]]
-    primer_matches: Dict[str, Dict[str, List[Dict]]]
-    variants: Dict[str, Dict[str, List[Dict]]]
+    segments: list[str]
+    segments_ref_id: dict[str, dict[str, str]]
+    segments_ref_seq: dict[str, dict[str, str]]
+    depths: dict[str, dict[str, str]]
+    mosdepth_info: dict[str, dict[str, FluMosdepthDepthInfo]]
+    primer_matches: dict[str, dict[str, list[dict]]]
+    variants: dict[str, dict[str, list[dict]]]
 
 
 class DB(DBMixin):
@@ -34,8 +32,8 @@ class DB(DBMixin):
     """
 
     ref_seq: str
-    depths: Dict[str, str]
-    amplicon_depths: Dict[str, List[Dict]]
-    mosdepth_info: Dict[str, MosdepthDepthInfo]
-    variants: Dict[str, List[Dict]]
-    echart_features: List[EChartsFeature]
+    depths: dict[str, str]
+    amplicon_depths: dict[str, list[dict]]
+    mosdepth_info: dict[str, MosdepthDepthInfo]
+    variants: dict[str, list[dict]]
+    echart_features: list[EChartsFeature]