+++ wgscovplot/features.py
@@ -1,5 +1,4 @@
 from itertools import cycle
-from typing import List, Optional
 
 from pydantic import BaseModel
 
@@ -42,10 +41,10 @@ class EChartsFeature(BaseModel):
 
 
 def build_echarts_features_array(
-    gene_features: Optional[List[Feature]],
-    amplicon_features: Optional[List[Feature]],
-    fp: Optional[FeaturesProps] = None,
-) -> List[EChartsFeature]:
+    gene_features: list[Feature] | None,
+    amplicon_features: list[Feature] | None,
+    fp: FeaturesProps | None = None,
+) -> list[EChartsFeature]:
     """Build a list of gene/amplicon feature properties for ECharts
 
     Args:
@@ -56,7 +55,7 @@ def build_echarts_features_array(
     """
     if fp is None:
         fp = FeaturesProps()
-    out: List[EChartsFeature] = []
+    out: list[EChartsFeature] = []
     colour_cycle = cycle(palette)
     fcminus = FeatureCoords()
     fcplus = FeatureCoords()